Refactor dashboard page
